+++ kinto_megaphone/listeners.py
@@ -18,11 +18,7 @@ class CollectionTimestampListener(ListenerBase):
 
         bucket_id = event.payload['bucket_id']
         collection_id = event.payload['collection_id']
-        parent_id = utils.instance_uri(event.request, 'collection',
-                                       bucket_id=bucket_id,
-                                       id=collection_id)
-        storage = event.request.registry.storage
-        timestamp = storage.collection_timestamp('record', parent_id)
+        timestamp = event.payload['timestamp']
         etag = '"{}"'.format(timestamp)
         self.client.send_version(self.broadcaster_id,
                                  '{}_{}'.format(bucket_id, collection_id),