+++ CHANGELOG.rst
@@ -7,7 +7,7 @@ This document describes changes between each past release.
 0.1.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Update to match kinto 10.1.1 -- see Kinto/kinto#1770 (#2).
 
 
 0.1.0 (2018-09-12)